shifted the template validation tests into the integration tests dir this encourages me to run the regular builder tests locally again
