+++ src/tests/test_buildercore_cfngen.py
@@ -1,4 +1,3 @@
-from time import sleep
 from . import base
 from buildercore import cfngen, project, context_handler
 
@@ -10,27 +9,6 @@ class TestBuildercoreCfngen(base.BaseCase):
         for pname in project.aws_projects().keys():
             LOG.info('rendering %s', pname)
             cfngen.quick_render(pname)
-
-    def test_validation(self):
-        "dummy projects and their alternative configurations pass validation"
-        for pname in project.aws_projects().keys():
-            self.assertTrue(cfngen.validate_project(pname))
-            sleep(0.25)
-
-    def test_validation_elife_projects(self):
-        "elife projects (and their alternative configurations) that come with the builder pass validation"
-
-        # HERE BE DRAGONS
-        # resets the testing config.SETTINGS_FILE we set in the base.BaseCase class
-        self.switch_out_test_settings()
-
-        for pname in project.aws_projects().keys():
-            with self.subTest(pname):
-                self.assertTrue(cfngen.validate_project(pname))
-                sleep(0.5)
-
-        # todo: does this need to live in a try: ... finally: ... ?
-        self.switch_in_test_settings()
 
 class TestBuildContext(base.BaseCase):
     def test_existing_alt_config(self):
